[CI/Build][Docs][Router] drop local pre-push lint hook
